backend/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudinary==1.40.0
cryptography==42.0.7
Django==4.2.13
django-cloudinary-storage==0.3.0
django-cors-headers==4.3.1
django-daraja==1.3.0
django-sendgrid-v5==1.2.3
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.7
mailchimp-transactional==1.0.56
packaging==24.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
requests==2.32.2
sendgrid==6.11.0
six==1.16.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
stripe==5.5.0
typing_extensions==4.11.0
urllib3==2.2.1
whitenoise==6.6.0
